+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './Text';
 export * from './CustomEditor';
+export * from './ResourcesEditor';
 export * from './TextPanel';