+++ src/constants/tests.ts
@@ -6,12 +6,6 @@ export const TestIds = {
     root: 'data-testid panel',
     fieldFrame: 'data-testid panel field-frame',
   },
-  helpersEditor: {
-    root: 'data-testid helpers-editor',
-  },
-  stylesEditor: {
-    root: 'data-testid styles-editor',
-  },
   text: {
     content: 'data-testid text content',
     error: 'data-testid text error',
@@ -19,5 +13,14 @@ export const TestIds = {
   },
   textEditor: {
     root: 'data-testid text-editor',
+  },
+  resourcesEditor: {
+    buttonAddNew: 'data-testid resources-editor button-add-new',
+    buttonRemove: 'data-testid resources-editor button-remove',
+    itemLabel: (name: string) => `data-testid resources-editor item-label-${name}`,
+    itemContent: (name: string) => `data-testid resources-editor item-content-${name}`,
+    fieldUrl: 'data-testid resources-editor field-url',
+    newItem: 'data-testid resources-editor new-item',
+    newItemName: 'data-testid resources-editor new-item-name',
   },
 };