+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './editor';
 export * from './panel';
+export * from './resource';